+++ app/[username]/campaigns/[id]/page.tsx
@@ -1,7 +1,7 @@
 import CampaignCard from '@/components/cards/CampaignCard';
 import {db} from '@/database/drizzle';
 import {campaigns} from '@/database/drizzle/schema';
-import {getUserFromSession} from '@/utils/supabase/helpers';
+import {getUserFromSession} from '@/server-actions/auth';
 import {redirect} from 'next/navigation';
 import {eq} from 'drizzle-orm';
