+++ components/buttons/AuthButton.tsx
@@ -2,7 +2,7 @@ import {createClient} from '@/utils/supabase/defaults/server';
 import Link from 'next/link';
 import {cookies} from 'next/headers';
 import LoginForm from '../forms/LoginForm';
-import {logoutAction} from '@/app/auth/actions';
+import {logoutAction} from '@/server-actions/auth';
 
 /**
  * Login/logout button depending on user auth state
